Separate locus and identifier into their own directories.
