+++ dart_bio_seqio/lib/src/seqio/infrastructure/genbank/locus/locus_dto.dart
@@ -1,4 +1,4 @@
-import '../../domain/entities/genbank/locus.dart';
+import '../../../domain/entities/genbank/locus.dart';
 
 class LocusDto {
   Locus fromGenbankFile({
