+++ dart_bio_seqio/test/src/seqio/infrastructure/genbank/feature/feature_dto_test.dart
@@ -1,9 +1,9 @@
 import 'package:dart_bio_seqio/src/seqio/domain/entities/genbank/location_position.dart';
-import 'package:dart_bio_seqio/src/seqio/infrastructure/genbank/feature_dto.dart';
+import 'package:dart_bio_seqio/src/seqio/infrastructure/genbank/feature/feature_dto.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../../data/genbank/SCU49845/SCU49845_genbank_original_format_data.dart';
-import '../../../../data/genbank/SCU49845/scu49845_genbank_data.dart';
+import '../../../../../data/genbank/SCU49845/SCU49845_genbank_original_format_data.dart';
+import '../../../../../data/genbank/SCU49845/scu49845_genbank_data.dart';
 
 void main() {
   late FeatureDto featuresDto;
